be/requirements.txt
pydantic
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
django
django-extensions
psycopg2-binary
pydantic[email]
python-multipart
httpx
python-dotenv